Prog2HF/HF11/requirements.txt
mypy==1.7.1
Pillow==10.1.0
requests==2.31.0
types-Pillow==10.1.0.2
types-requests==2.31.0.10
